Adding more Ui tests
